@@ -30,39 +30,39 @@
   },
   "description": "Get position and size of the DOM element for any React Component",
   "devDependencies": {
-    "ava": "^0.18.1",
-    "babel": "6.5.2",
-    "babel-cli": "6.18.0",
+    "ava": "^0.18.2",
+    "babel": "^6.23.0",
+    "babel-cli": "^6.23.0",
     "babel-eslint": "7.1.1",
-    "babel-loader": "6.2.10",
+    "babel-loader": "^6.3.2",
     "babel-plugin-add-module-exports": "0.2.1",
     "babel-plugin-transform-decorators-legacy": "1.3.4",
-    "babel-polyfill": "^6.22.0",
-    "babel-preset-env": "^1.1.8",
-    "babel-preset-react": "^6.22.0",
+    "babel-polyfill": "^6.23.0",
+    "babel-preset-latest": "^6.22.0",
+    "babel-preset-react": "^6.23.0",
     "babel-preset-stage-2": "^6.22.0",
-    "eslint": "^3.15.0",
-    "eslint-config-rapid7": "^2.7.0",
+    "browser-env": "^2.0.21",
+    "eslint": "^3.16.1",
+    "eslint-config-rapid7": "^2.8.1",
     "eslint-friendly-formatter": "2.0.7",
-    "eslint-loader": "1.6.1",
-    "eslint-plugin-react": "6.9.0",
-    "file-loader": "^0.10.0",
+    "eslint-loader": "^1.6.3",
+    "eslint-plugin-react": "^6.10.0",
+    "file-loader": "^0.10.1",
     "font-awesome": "4.7.0",
     "html-webpack-plugin": "^2.28.0",
     "in-publish": "2.0.0",
     "ink-docstrap": "1.3.0",
     "jsdoc": "3.4.3",
     "jsdoc-babel": "0.3.0",
-    "jsdom": "^9.10.0",
-    "lodash-webpack-plugin": "^0.11.0",
+    "lodash-webpack-plugin": "^0.11.2",
     "nyc": "^10.1.2",
     "optimize-js-plugin": "0.0.4",
     "react-dom": "15.4.2",
-    "rimraf": "2.5.4",
+    "rimraf": "^2.6.1",
     "sinon": "1.17.7",
-    "webpack": "1.14.0",
+    "webpack": "^2.2.1",
     "webpack-dashboard": "^0.3.0",
-    "webpack-dev-server": "1.16.2"
+    "webpack-dev-server": "^2.4.1"
   },
   "homepage": "https://github.com/planttheidea/remeasure#readme",
   "keywords": [
@@ -96,12 +96,12 @@
     "docs": "jsdoc -c ./jsdoc.config.json && git add --all",
     "lint": "NODE_ENV=test eslint src",
     "prepublish": "in-publish && npm run prepublish:compile || echo ''",
-    "prepublish:compile": "npm run lint && npm run clean && npm run test:coverage && npm run transpile && npm run dist && npm run docs",
+    "prepublish:compile": "npm run lint && npm run test:coverage && npm run clean && npm run transpile && npm run dist && npm run docs",
     "start": "npm run dev",
     "test": "NODE_ENV=test ava --no-cache",
     "test:coverage": "nyc npm test",
     "test:watch": "NODE_ENV=test ava --watch",
     "transpile": "babel src --out-dir lib"
   },
-  "version": "2.1.3"
+  "version": "2.1.4"
 }
