@@ -8,8 +8,6 @@ const defaultConfig = require('./webpack.config');
 module.exports = Object.assign({}, defaultConfig, {
   cache: false,
 
-  debug: false,
-
   devtool: undefined,
 
   output: Object.assign({}, defaultConfig.output, {
@@ -17,15 +15,23 @@ module.exports = Object.assign({}, defaultConfig, {
   }),
 
   plugins: defaultConfig.plugins.concat([
+    new webpack.LoaderOptionsPlugin({
+      debug: false,
+      minimize: true
+    }),
+    new webpack.optimize.OccurrenceOrderPlugin(),
     new webpack.optimize.UglifyJsPlugin({
       compress: {
         booleans: true,
         conditionals: true,
+        dead_code: true,
         drop_console: true,
         drop_debugger: true,
+        evaluate: true,
         join_vars: true,
         screw_ie8: true,
         sequences: true,
+        unused: true,
         warnings: false
       },
       sourceMap: false
